requirements.txt
torch>=2.1.0
transformers>=4.40.0
vllm>=0.4.0
accelerate>=0.28.0
Pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
pyyaml>=6.0
cairosvg>=2.7.0
graphviz>=0.20.0
pdf2image>=1.16.0
pytorch-fid>=0.3.0
qwen-vl-utils>=0.0.2
tqdm>=4.65.0
wandb>=0.16.0
